% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moody.R
\name{Moody}
\alias{Moody}
\title{Pick a mood}
\usage{
Moody()
}
\value{
Random selected song from list
}
\description{
Asks the user for an input from (R&B/Soul, Alternative/Indie, Hip-Hop/Rap, Pop, Dance/Electronic)
}
